backend/apps/llm-python-rpc/requirements.txt
grpcio>=1.75.1
grpcio-tools>=1.75.1
protobuf>=5.0.0
requests>=2.31.0
pypdf>=5.0.0
